Committing Board.java
I am committing my Board class, which contains all my main code for initialising a game of chess. It maps tile co-ordinates on a chess board to the starting positions of each piece, as well as mapping each tile co-ordinate to their status (such as whether a tile is starting as Occupied or Empty).