Created Piece Class
Created a Piece class. This contains a boolean method for checking/validating that the player has selected to move a valid piece (their own piece). It contains a piece type method to check the type of piece the player is moving based on the type inside the PieceID. Finally, it has an 'isOpponentPiece' method, which checks to see if the tile the player is moving to has an opponent piece inside or not (which they can then capture).