Added setPos Method to Rook class
Added a new setPos method in the Rook class, which is responsible for finalising a move/setting the new position of a Rook. It takes into account opponent pieces in its destination tile to capture that piece, or doesn't move the Rook if there is an own player piece in the destination tile.

Also added or improved general Javadoc comments to the class.