Fixed Bug in Rook Class
Fixed a bug in the Rook class, where I was checking for null in my if statements but passing "null" (as a String). However, I cannot change "null" to null for my if statements as TreeMaps do not accept null values with the way I have set it up. Therefore, I changed my if statement checks to check for the tile occupation status instead.