Fixed Bug in Queen Class
Fixed Bug in Queen class relating to Diagonal movements, where having a nested for loop would iterate through all the columns each time the outer loop (for the rows) iterates. This was not first realised and isn't the behaviour intended for the class. Therefore used an int j to increment/decrement (according to the column moving left or right) each time the row for loop iterates.